backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
supabase==2.15.1
python-dotenv==1.1.0
PyJWT[crypto]==2.10.1
requests==2.32.3